+++ src/features/PopularRoutes/ui/PopularRoutes/PopularRoutes.stories.tsx
@@ -1,5 +1,5 @@
 import type { Meta, StoryObj } from '@storybook/react-vite';
-import { PopularRoutes } from 'src/features/PopularRoutes';
+import { PopularRoutes } from './PopularRoutes';
 
 const meta = {
   title: 'Features/PopularRoutes',
