+++ src/features/SpotPreview/ui/CitySelector/CitySelector.tsx
@@ -1,4 +1,4 @@
-import type { CitySelectorProps } from '../model/types';
+import type { CitySelectorProps } from '../../model/types';
 
 export function CitySelector({ cities, selected, onSelect }: CitySelectorProps) {
   return (
